requirements.txt
Flask>=3.0.0
PyPDF2>=3.0.1
sentence-transformers>=2.2.2
faiss-cpu==1.12.0
torch>=2.0.0
transformers>=4.41.0
tokenizers>=0.15.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.65.0
langchain>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
